refactor(console): split module catalog out of types barrel and flatten remaining importer hotspots
