+++ console/shared/lib/types.ts
@@ -2,6 +2,12 @@
 // Proprietary. No license granted. See LICENSE.md.
 
 import type { IconName } from './icons';
+import { MODULE_CATALOG, type ModuleDef } from './module-catalog';
+
+// The built-in command + module catalogs live beside this barrel in
+// ./module-catalog and flow through here unchanged.
+export * from './module-catalog';
+
 // Wire types mirroring the Go NATS RPC contracts (JSON over core NATS).
 export type Perm = 'everyone' | 'sub' | 'vip' | 'mod' | 'lead_mod' | 'broadcaster';
 export type Tier = 'premium' | 'standard';
@@ -40,114 +46,6 @@ export interface CommandView {
   builtin?: boolean;
 }
 
-// --- Built-in command catalog --------------------------------------------
-// Built-in commands are behaviors baked into the bot (not user text). They show
-// on the commands page alongside custom commands, flagged builtin, but they
-// cannot be renamed, deleted, or given a custom response — only toggled on/off.
-// Their per-user on/off state lives in the modules service under `id` (a missing
-// row means defaultActive). Adding one is a row here + the matching sesame
-// built-in module. They are deliberately NOT in MODULE_CATALOG (never shown on
-// the modules page).
-
-export interface BuiltinCommandDef {
-  // id is both the chat trigger and the modules-service key for the toggle.
-  id: string;
-  label: string;
-  // summary is shown in the command row where a custom command shows its
-  // response (built-ins have no response).
-  summary: string;
-  description: string; // longer copy for the inspector
-  // usage lists example invocations shown in the inspector.
-  usage: string[];
-  // preview is the bot REPLY template, rendered through ChatPreview (as a
-  // reply rehearsal: only previewSamples substitute — built-in replies are
-  // bare token replacers with no dynamic tokens or slash-verb routing).
-  // previewArgs is what the viewer types after the trigger.
-  preview: string;
-  previewArgs?: string;
-  previewSamples?: Record<string, string>;
-  defaultActive: boolean;
-  defaultPerm: Perm;
-  defaultCooldown: number; // seconds
-  // liveOnly commands run only while the broadcaster is streaming.
-  liveOnly: boolean;
-  // editable: the reply template can be customized on the dashboard. When true
-  // the inspector shows a ResponseEditor (with the `tokens` palette) and a
-  // rehearsal, and saves the template into the modules-service config under
-  // `replyKey`. The bot expands the tokens when it posts the reply (e.g. {clip}
-  // → the clip URL, resolved by outgress once the clip exists). Non-editable
-  // built-ins stay a read-only preview. `preview` doubles as the default
-  // template when no custom reply is set.
-  editable?: boolean;
-  // replyKey is the Configs key the custom reply template is stored under (only
-  // meaningful when editable).
-  replyKey?: string;
-  // tokens is the reply editor's insert palette (token names without braces).
-  tokens?: string[];
-}
-
-export const BUILTIN_COMMANDS: readonly BuiltinCommandDef[] = [
-  {
-    id: 'accountage',
-    label: 'Account age',
-    summary: 'Built-in · shows how long a Twitch account has existed.',
-    description:
-      'Shows how old your Twitch account is, or the age of another Twitch user\'s account when you add their username.',
-    usage: ['!accountage', '!accountage <user>'],
-    preview: "@{target}'s account is 4 years, 2 months old.",
-    previewArgs: 'viewer',
-    previewSamples: { target: 'viewer' },
-    defaultActive: true,
-    defaultPerm: 'everyone',
-    defaultCooldown: 15,
-    liveOnly: false
-  },
-  {
-    id: 'followage',
-    label: 'Followage',
-    summary: 'Built-in · shows how long a viewer has followed the channel.',
-    description:
-      'Shows how long you or another Twitch user has followed the channel. Add a username to look up someone else.',
-    usage: ['!followage', '!followage <user>'],
-    preview: '@{target} has followed for 8 months.',
-    previewArgs: 'viewer',
-    previewSamples: { target: 'viewer' },
-    defaultActive: true,
-    defaultPerm: 'everyone',
-    defaultCooldown: 15,
-    liveOnly: false
-  },
-  {
-    id: 'clip',
-    label: 'Clip',
-    summary: 'Built-in · clips the last moments of the stream and posts the link.',
-    description:
-      'Viewers create a clip of the recent stream and the bot replies in chat with the clip link. Add an optional title after the command. Only works while you are live.',
-    usage: ['!clip', '!clip <title>'],
-    // Real reply format: "<clipper> clipped: <title> → <url>" (see
-    // app/outgress/internal/worker clipReplyText). {user} = the clipper, {target}
-    // = the title argument (standard command token).
-    preview: '{user} clipped: {target} → {clip}',
-    previewArgs: 'That is amazing',
-    previewSamples: { user: 'sesame_sam', target: 'That is amazing', clip: 'clips.twitch.tv/AbCdEf' },
-    defaultActive: true,
-    defaultPerm: 'everyone',
-    defaultCooldown: 15,
-    liveOnly: true,
-    // The reply is customizable: {clip} is the clip link, {user} the clipper,
-    // {target} the title the viewer typed. Stored under the "reply" config key,
-    // read by sesame and expanded by outgress (see app/sesame/modules/clip.go).
-    editable: true,
-    replyKey: 'reply',
-    tokens: ['clip', 'user', 'target']
-  }
-];
-
-export function builtinDef(id: string): BuiltinCommandDef | undefined {
-  return BUILTIN_COMMANDS.find((b) => b.id === id);
-}
-
-export const BUILTIN_NAMES: ReadonlySet<string> = new Set(BUILTIN_COMMANDS.map((b) => b.id));
 
 export interface AdminUser {
   user_id: string;
@@ -247,1538 +145,6 @@ export interface DashboardLink {
   name: string;
 }
 
-// --- Module catalog -------------------------------------------------------
-// The user-facing modules the dashboard lets a broadcaster toggle/configure.
-// Core, hidden modules (the command processor, the live tracker, and the system
-// module that owns !ping/!itsbagelbot and the bagel greeting) are deliberately
-// NOT listed here: they are always on and never shown.
-
-export type ModuleFieldType = 'text' | 'textarea' | 'number' | 'select' | 'toggle' | 'timezone';
-
-export interface ModuleField {
-  // key is the JSON property written into the module's Configs blob.
-  key: string;
-  label: string;
-  type: ModuleFieldType;
-  placeholder?: string;
-  help?: string;
-  // options drive a 'select' field.
-  options?: { value: string; label: string }[];
-  // followsLevel marks a 'toggle' whose unset state follows the module's
-  // "level" select (see automodToggleDefault): the blob only stores an
-  // explicit "on"/"off" once the user flips it.
-  followsLevel?: boolean;
-}
-
-// Mirrors levelSections in app/sesame/automod/config.go: which automod sections
-// each level preset enables. Renders the resting state of a follows-level
-// toggle; the authoritative resolution happens in Go.
-export const AUTOMOD_LEVEL_DEFAULTS: Record<string, Record<string, boolean>> = {
-  none: { harassment: false, sexual: false, profanity: false, style: false, links: false },
-  basic: { harassment: true, sexual: false, profanity: false, style: false, links: false },
-  moderate: { harassment: true, sexual: true, profanity: false, style: true, links: true },
-  strict: { harassment: true, sexual: true, profanity: true, style: true, links: true }
-};
-
-// automodToggleDefault resolves a follows-level toggle's resting state for the
-// currently selected level.
-export function automodToggleDefault(level: string, key: string): boolean {
-  return (AUTOMOD_LEVEL_DEFAULTS[level] ?? AUTOMOD_LEVEL_DEFAULTS.moderate)[key] ?? false;
-}
-
-// One chat line a module can post, rendered as a row on the module page. Clicking
-// the row opens the exact same builder as a custom command's response (the shared
-// ResponseEditor + ChatPreview, standard {user}/{target}/… tokens). messageKey/
-// enableKey are the Configs JSON keys the matching sesame module reads (see
-// app/sesame/modules).
-export interface ModuleReply {
-  key: string; // stable row id
-  label: string; // 'Follow alert'
-  tagline: string; // short row description
-  // Preview context: what fires this line, e.g. 'on follow'.
-  event: string;
-  messageKey: string; // Configs key holding the template
-  // Configs key for this reply's own on/off toggle; omit when the reply has no
-  // per-reply switch (it fires whenever the module is on). Stored "on"/"off";
-  // empty/absent means on, matching sesame's alertOn semantics.
-  enableKey?: string;
-  // Flips the empty/absent default to OFF: the reply fires only on an explicit
-  // "on" (sesame's adAlertOn semantics). Used by alerts the broadcaster must
-  // opt into, like the ad-break announcement.
-  defaultOff?: boolean;
-  defaultMessage: string; // sesame default template (placeholder + preview fallback)
-
-  // --- command-style replies (gossip modules: urchin, mcsr) ---------------
-  // command is the chat trigger without '!' (e.g. 'daily'). When set, the
-  // inspector rehearses the reply exactly like a custom command: the border
-  // reads "Chat rehearsal", a viewer line types the trigger, and the bot
-  // answers with previewSamples substituted into the template.
-  command?: string;
-  // previewArgs is what the sample viewer types after the trigger.
-  previewArgs?: string;
-  // previewSamples maps this reply's tokens to sample values. The rehearsal
-  // (kind="reply") substitutes ONLY these plus the shared dynamic tokens:
-  // sesame expands only the module's own token map here, so the generic
-  // command samples would preview values the bot will never produce.
-  previewSamples?: Record<string, string>;
-  // tokens is the editor's insert palette (without braces), replacing the
-  // default command tokens with the ones this reply actually supports.
-  tokens?: string[];
-}
-
-// A chat command a module exposes, listed read-only on the module page so a
-// broadcaster can see what turning the module on unlocks. Unlike a ModuleReply
-// these are not editable or toggleable: modules whose replies are fixed system
-// lines (the queue) have nothing to configure per command, so the rows are
-// informational only — never clickable.
-export interface ModuleCommandInfo {
-  trigger: string; // the chat trigger with '!' (e.g. '!join', '!queue next')
-  summary: string; // one-line description of what it does
-  // perm names the minimum role, shown as a small tag. Omit for everyone.
-  perm?: 'mod';
-}
-
-export interface ModuleDef {
-  // id is normally the ModuleView.name key in the modules service. Catalog
-  // tools with toggleable=false use it only as their stable UI key.
-  id: string;
-  label: string;
-  tagline: string; // one-liner for the tile
-  description: string; // longer copy for the module page
-  icon: IconName;
-  category: string;
-  defaultEnabled: boolean;
-  // False for catalog tools that live under Modules but are always available
-  // rather than backed by an enableable row in the modules service. They keep
-  // the shared tile/delegation model without presenting a meaningless switch.
-  toggleable?: boolean;
-  // If true, the module is hidden from the dashboard and unreachable.
-  hidden?: boolean;
-  // The module's configurable chat lines (the "commands" of the module page).
-  replies: ModuleReply[];
-  // Read-only chat commands to list on the module page. For modules that expose
-  // commands with fixed (non-customizable) replies, e.g. the play queue. Shown
-  // as static rows, never clickable. Optional.
-  commands?: ModuleCommandInfo[];
-  // Plain non-reply settings (rendered in the settings strip). Optional; the
-  // current modules have none beyond their master enable + per-reply toggles.
-  settings?: ModuleField[];
-  // href overrides the tile's link when a module needs a bespoke inspector
-  // instead of the generic /modules/[id] reply page (govee's device + reward
-  // setup). Absent for the ordinary reply-configured modules.
-  href?: string;
-  // delegateSections names the delegation grants that open the bespoke href
-  // page for a delegate. Defaults to ['modules'] (the tile lives on /modules),
-  // so it only needs setting when a page has its own grant (channel points) or
-  // additionally rides another one (timers/quotes under 'commands'). Read via
-  // moduleDelegateSections; the route guard, the per-page gates and the tile
-  // grid all derive from it, so a new bespoke page never touches those.
-  delegateSections?: readonly string[];
-}
-
-// moduleDelegateSections resolves a def's delegation scope (see the field doc).
-export function moduleDelegateSections(def: ModuleDef): readonly string[] {
-  return def.delegateSections ?? ['modules'];
-}
-
-// A module's current state as shown on the dashboard: catalog metadata merged
-// with the broadcaster's stored row.
-export interface ModuleState {
-  def: ModuleDef;
-  enabled: boolean;
-  config: Record<string, string>;
-}
-
-// Shared token palette + preview samples for the Bedwars session commands
-// (!daily / !weekly / !monthly) — same template surface, one source of truth.
-const BW_SESSION_TOKENS = [
-  'player',
-  'wins',
-  'losses',
-  'finals',
-  'finaldeaths',
-  'beds',
-  'games',
-  'levels',
-  'fkdr'
-];
-const BW_SESSION_SAMPLES: Record<string, string> = {
-  player: 'Technoblade',
-  wins: '5',
-  losses: '2',
-  finals: '21',
-  finaldeaths: '3',
-  beds: '9',
-  games: '8',
-  levels: '1',
-  fkdr: '7.00'
-};
-
-// Shared token palette + preview samples for the Fortnite stats commands
-// (!fnstats / !season) — same template surface, one source of truth.
-const FN_STATS_TOKENS = [
-  'player',
-  'window',
-  'wins',
-  'matches',
-  'kills',
-  'kd',
-  'winrate',
-  'solowins',
-  'solomatches',
-  'solokd',
-  'duowins',
-  'duomatches',
-  'duokd',
-  'squadwins',
-  'squadmatches',
-  'squadkd'
-];
-const FN_STATS_SAMPLES: Record<string, string> = {
-  player: 'Ninja',
-  window: 'lifetime',
-  wins: '301',
-  matches: '6232',
-  kills: '21679',
-  kd: '3.66',
-  winrate: '4.83',
-  solowins: '120',
-  solomatches: '2400',
-  solokd: '3.2',
-  duowins: '90',
-  duomatches: '1900',
-  duokd: '3.8',
-  squadwins: '91',
-  squadmatches: '1932',
-  squadkd: '4.1'
-};
-
-// Token palette + preview samples for !fn session: deltas since the
-// stream-start snapshot (no per-mode breakdown, no window — always this
-// stream, always the linked account).
-const FN_SESSION_TOKENS = ['player', 'wins', 'matches', 'kills', 'kd', 'winrate'];
-const FN_SESSION_SAMPLES: Record<string, string> = {
-  player: 'Ninja',
-  wins: '3',
-  matches: '12',
-  kills: '48',
-  kd: '5.33',
-  winrate: '25.0'
-};
-
-export const MODULE_CATALOG: readonly ModuleDef[] = [
-  // Chat Tools: the bot's viewer-facing chat features, surfaced first. Channel
-  // Points and Timers own bespoke pages (opened via href); Trigger Words uses
-  // the generic reply inspector with its rule editor.
-  {
-    id: 'channelpoints',
-    label: 'Channel Points',
-    tagline: 'Turn channel-point redemptions into bot actions.',
-    description:
-      'Create the custom rewards viewers redeem with channel points (made under the bot on Twitch, styled natively) and bind each one to a bot action, like posting a chat line. Choose whether each redemption is fulfilled, refunded, or left for a mod. Manage the rewards on this page.',
-    icon: 'gem',
-    category: 'Chat Tools',
-    defaultEnabled: false,
-    href: '/channelpoints',
-    // Channel Points is its own delegation grant (see SECTIONS in the settings
-    // page), not part of the blanket 'modules' one.
-    delegateSections: ['channelpoints'],
-    replies: []
-  },
-  {
-    id: 'timers',
-    label: 'Timers',
-    tagline: 'Post repeating chat messages on a schedule while you are live.',
-    description:
-      'Set messages the bot repeats on a schedule while you are live: announcements, socials, reminders. Each timer keeps its own interval and only fires during the stream. Add, edit and arm them on this page.',
-    icon: 'clock',
-    category: 'Chat Tools',
-    defaultEnabled: false,
-    href: '/timers',
-    // Timers has no standalone grant: the 'commands' grant has always covered
-    // it too (see the settings page SECTIONS comment). 'timers' keeps legacy
-    // grants minted with a bare timers section working.
-    delegateSections: ['modules', 'commands', 'timers'],
-    replies: []
-  },
-  {
-    id: 'loyalty',
-    label: 'Loyalty Points',
-    tagline: 'Viewers earn channel currency for subs, cheers and watch time.',
-    description:
-      'Give your community its own currency: viewers earn points for subscribing, resubscribing, gifting subs, cheering bits, and simply watching (everyone in chat earns on a 5-minute tick while you are live). Name the currency, tune every rate, and let mods grant points with !points set/add. Viewers check their standing with !points. Pairs with Counters and channel-point rewards that award points.',
-    icon: 'coin',
-    category: 'Community',
-    defaultEnabled: false,
-    href: '/loyalty',
-    replies: []
-  },
-  {
-    id: 'counters',
-    label: 'Counters',
-    tagline: 'Track wins, deaths, hugs, redeems — anything your chat can count.',
-    description:
-      'Create channel-wide, per-command, or per-viewer tallies. Adjust them from the dashboard or chat, and bump them from command responses and channel-point rewards.',
-    icon: 'list',
-    category: 'Community',
-    defaultEnabled: true,
-    toggleable: false,
-    href: '/counters',
-    // The default Modules delegation scope gives delegates access to the full
-    // counter book; commands-only delegates retain picker access separately.
-    replies: []
-  },
-  {
-    id: 'gamble',
-    label: 'Gamble',
-    tagline: 'Let viewers wager their points on a roll with !gamble.',
-    description:
-      'Give your loyalty points a game: viewers type !gamble <amount> (or half/all of their standing) and the bot rolls 1-100. Landing inside your win chance pays the stake back plus its match; anything else takes it. Set the win odds, bet limits and per-viewer cooldown below, rename the currency in the replies, and customize the win/lose lines. Every payout and debit moves real loyalty points through the same ledger as !points.',
-    icon: 'gamepad',
-    category: 'Community',
-    defaultEnabled: false,
-    // The numeric knobs are plain settings the generic page patches into the
-    // module blob; sesame clamps them server-side (engine.ClampGambleSettings),
-    // so an out-of-range save can never arm an unlimited machine.
-    settings: [
-      { key: 'pointsName', label: 'Currency name', type: 'text', placeholder: 'points', help: 'The word used in the win/lose lines (defaults to "points").' },
-      { key: 'winPercent', label: 'Win chance %', type: 'number', placeholder: '50', help: 'A roll of this number or lower wins. 50 is a fair coin; 1-99 allowed.' },
-      { key: 'minBet', label: 'Minimum bet', type: 'number', placeholder: '1' },
-      { key: 'maxBet', label: 'Maximum bet', type: 'number', placeholder: '1000' },
-      { key: 'cooldownSeconds', label: 'Cooldown (seconds)', type: 'number', placeholder: '10', help: 'Per viewer — one chatter gambling never blocks another.' }
-    ],
-    replies: [
-      {
-        key: 'won',
-        label: 'Win line',
-        tagline: 'When a roll lands inside the win chance.',
-        event: '!gamble 100',
-        command: 'gamble',
-        previewArgs: '100',
-        messageKey: 'winMessage',
-        defaultMessage: '@{user} rolled {roll} (needed {chance} or less) and won {amount} {points} — now at {balance}!',
-        tokens: ['user', 'roll', 'chance', 'amount', 'balance', 'points'],
-        previewSamples: { user: 'sesame_sam', roll: '23', chance: '50', amount: '100', balance: '1340', points: 'points' }
-      },
-      {
-        key: 'lost',
-        label: 'Loss line',
-        tagline: 'When a roll misses the win chance.',
-        event: '!gamble 100',
-        command: 'gamble',
-        previewArgs: '100',
-        messageKey: 'loseMessage',
-        defaultMessage: '@{user} rolled {roll} (needed {chance} or less) and lost {amount} {points}. Now at {balance}.',
-        tokens: ['user', 'roll', 'chance', 'amount', 'balance', 'points'],
-        previewSamples: { user: 'sesame_sam', roll: '87', chance: '50', amount: '100', balance: '1140', points: 'points' }
-      }
-    ],
-    commands: [
-      { trigger: '!gamble <amount>', summary: 'Wager that many points on a roll.' },
-      { trigger: '!gamble half|all', summary: 'Wager half of / the whole standing.' }
-    ]
-  },
-  {
-    id: 'duel',
-    label: 'Duels',
-    tagline: 'Viewer-vs-viewer point duels: pot free-for-alls and 1v1 challenges.',
-    description:
-      "Two ways to duel for points. A pot duel: someone types !duel <stake> and everyone has the window to add their own stake — when time runs out the bot draws one winner weighted by stake and they take the whole pot. Or a challenge: !duel <user> <stake> names an opponent who must type !duel accept before the window closes; equal stakes, a fair coin flip, winner takes both. Decline, cancellation and no-shows always refund every escrowed point, and every movement goes through the loyalty service's guarded spend — nobody can wager what they do not have.",
-    icon: 'activity',
-    category: 'Community',
-    defaultEnabled: false,
-    settings: [
-      { key: 'pointsName', label: 'Currency name', type: 'text', placeholder: 'points', help: 'The word used in money lines (defaults to "points").' },
-      { key: 'minStake', label: 'Minimum stake', type: 'number', placeholder: '1' },
-      { key: 'maxStake', label: 'Maximum stake', type: 'number', placeholder: '1000' },
-      { key: 'potSeconds', label: 'Pot window (seconds)', type: 'number', placeholder: '60', help: 'How long a pot duel accepts stakes.' },
-      { key: 'challengeSeconds', label: 'Accept window (seconds)', type: 'number', placeholder: '120', help: 'How long the challenged party has to accept.' }
-    ],
-    replies: [
-      {
-        key: 'opened',
-        label: 'Pot duel opened',
-        tagline: 'When a pot duel starts and stakes open.',
-        event: '!duel 100',
-        command: 'duel',
-        previewArgs: '100',
-        messageKey: 'openedMessage',
-        defaultMessage: 'Pot duel is LIVE! @{user} put up {stake} {points} — type !duel <amount> to join. Drawing in {secs}s!',
-        tokens: ['user', 'stake', 'secs', 'points'],
-        previewSamples: { user: 'sesame_sam', stake: '100', secs: '60', points: 'points' }
-      },
-      {
-        key: 'joined',
-        label: 'Join confirmation',
-        tagline: 'When a viewer adds their stake to a running pot.',
-        event: '!duel 250',
-        command: 'duel',
-        previewArgs: '250',
-        messageKey: 'joinMessage',
-        defaultMessage: "@{user} you're in with {stake}! {count} in the duel, {pot} {points} in the pot.",
-        tokens: ['user', 'stake', 'count', 'pot', 'points'],
-        previewSamples: { user: 'sesame_sam', stake: '250', count: '4', pot: '700', points: 'points' }
-      },
-      {
-        key: 'challenge',
-        label: 'Challenge sent',
-        tagline: 'When a viewer challenges another to even stakes.',
-        event: '!duel @maya_live 500',
-        command: 'duel',
-        previewArgs: '@maya_live 500',
-        messageKey: 'challengeMessage',
-        defaultMessage: '@{user} challenges @{target} for {stake} {points}! @{target}, type !duel accept within {secs}s — winner takes {pot}!',
-        tokens: ['user', 'target', 'stake', 'pot', 'secs', 'points'],
-        previewSamples: { user: 'sesame_sam', target: 'maya_live', stake: '500', pot: '1000', secs: '120', points: 'points' }
-      },
-      {
-        key: 'won',
-        label: 'Result line',
-        tagline: 'When a challenge is accepted and settled.',
-        event: '!duel accept',
-        command: 'duel',
-        previewArgs: 'accept',
-        messageKey: 'wonMessage',
-        defaultMessage: 'The blades fall — @{winner} defeats @{loser} and takes {pot} {points}!',
-        tokens: ['winner', 'loser', 'pot', 'points'],
-        previewSamples: { winner: 'maya_live', loser: 'sesame_sam', pot: '1000', points: 'points' }
-      }
-    ],
-    commands: [
-      { trigger: '!duel', summary: 'Show what runs: a pot (entrants, pool, seconds) or a pending challenge.' },
-      { trigger: '!duel <stake>', summary: 'Open a pot duel — or join one that is running.' },
-      { trigger: '!duel <user> <stake>', summary: 'Challenge someone to even stakes; winner takes both.' },
-      { trigger: '!duel accept', summary: 'The challenged party matches the stake and settles it instantly.' },
-      { trigger: '!duel decline', summary: 'Refuse a challenge; the opener is refunded.', perm: 'mod' },
-      { trigger: '!duel cancel', summary: 'Tear the running duel down with full refunds (opener or mod).', perm: 'mod' }
-    ]
-  },
-  {
-    id: 'triggers',
-    label: 'Trigger Words',
-    tagline: 'Auto-reply when a word shows up in chat — no "!" needed.',
-    description:
-      'Give the bot a list of words or phrases and the line to post when it sees one in ordinary chat — no command prefix required. Each rule pairs a phrase with a response: pick how the phrase matches (whole word, contains, exact message or starts with — the default matches the whole word, so "hi" will not fire inside "this"), write the reply, and switch individual rules on or off without deleting them. Responses support {user}, {random} and {choice:a,b,c}. The first matching rule wins, so one message gets at most one reply.',
-    icon: 'caps',
-    category: 'Chat Tools',
-    defaultEnabled: false,
-    // Trigger rules are a free-form list of phrase→response pairs the author grows,
-    // so the module page renders them as add/removable ReplyRows with a bespoke
-    // rule inspector (TriggerRuleEditor: phrase, match mode, response, per-rule
-    // enabled switch). The whole list is persisted as one "rules" string holding
-    // a JSON array of structured rules; the sesame module also still parses the
-    // legacy "phrase => response" line format for configs saved before the
-    // migration. See app/sesame/modules/triggers.go.
-    replies: []
-  },
-  {
-    id: 'time',
-    label: 'Local Time',
-    tagline: 'Viewers ask what time it is for you with !time.',
-    description:
-      'Viewers type !time and the bot answers with your current local time. Pick your timezone below — the page suggests the one your browser reports, computed on your device only (nothing is read or stored until you save it). Choose a 12- or 24-hour clock and customize the reply.',
-    icon: 'globe',
-    category: 'Chat Tools',
-    defaultEnabled: false,
-    replies: [
-      {
-        key: 'time',
-        label: '!time',
-        tagline: 'Your current local time.',
-        event: '!time',
-        command: 'time',
-        messageKey: 'message',
-        defaultMessage: 'It is currently {time} for the streamer.',
-        tokens: ['time', 'date', 'timezone', 'user'],
-        previewSamples: {
-          time: '2:30 PM',
-          date: 'Monday, July 13',
-          timezone: 'America/Toronto',
-          user: 'Viewer'
-        }
-      }
-    ],
-    settings: [
-      {
-        key: 'timezone',
-        label: 'Timezone',
-        type: 'timezone',
-        placeholder: 'America/Toronto',
-        help: 'IANA timezone name used to render {time} and {date}. The suggestion comes from your browser and is never stored until you save it.'
-      },
-      {
-        key: 'format',
-        label: 'Clock format',
-        type: 'select',
-        placeholder: '12',
-        options: [
-          { value: '12', label: '12-hour (2:30 PM)' },
-          { value: '24', label: '24-hour (14:30)' }
-        ]
-      }
-    ]
-  },
-  {
-    id: 'automod',
-    label: 'AutoMod',
-    tagline: 'Catch scams, IP-grabbers and raid spam before your mods do.',
-    description:
-      'The bot screens every chat line for harmful content and coordinated raid floods, and warns, deletes, times out or bans the sender. Trusted chatters (VIPs, mods, the broadcaster) are always exempt, and anything borderline is left to your human mods. Pick a level from None to All, then fine-tune each check below. The safety floor (hate slurs and IP-grabber links) is always enforced, on every level and even with the module off: hosting those risks your channel and the bot account platform-wide. Everything else is your call.',
-    icon: 'moderation',
-    category: 'Moderation',
-    defaultEnabled: true,
-    hidden: true,
-    // AutoMod is pure configuration: no chat reply lines, only the settings strip.
-    replies: [],
-    settings: [
-      {
-        key: 'level',
-        label: 'Enforcement level',
-        type: 'select',
-        placeholder: 'moderate',
-        options: [
-          { value: 'none', label: 'None - safety floor only' },
-          { value: 'basic', label: 'Basic - floor + harassment' },
-          { value: 'moderate', label: 'Moderate - recommended (default)' },
-          { value: 'strict', label: 'All - every check, family-strict' }
-        ],
-        help: 'Sets the default for every check below. The safety floor applies at every level.'
-      },
-      {
-        key: 'harassment',
-        label: 'Harassment',
-        type: 'toggle',
-        followsLevel: true,
-        help: 'Directed harm ("kys" and friends): warns the sender and removes the message; repeat offenders are timed out, then banned.'
-      },
-      {
-        key: 'sexual',
-        label: 'Sexual content',
-        type: 'toggle',
-        followsLevel: true,
-        help: 'Removes messages with explicit sexual terms.'
-      },
-      {
-        key: 'profanity',
-        label: 'Profanity',
-        type: 'toggle',
-        followsLevel: true,
-        help: 'Removes plain swearing. Off by default: most channels allow it.'
-      },
-      {
-        key: 'style',
-        label: 'Caps & symbol spam',
-        type: 'toggle',
-        followsLevel: true,
-        help: 'Removes shouting, symbol walls and character floods. Emote walls (KEKW spam) are recognized and never flagged.'
-      },
-      {
-        key: 'links',
-        label: 'Link-spam radar',
-        type: 'toggle',
-        followsLevel: true,
-        help: 'Watches for the same link template posted by many different accounts and removes the wave. Single links are never touched.'
-      },
-      {
-        key: 'block_terms',
-        label: 'Blocked terms',
-        type: 'textarea',
-        placeholder: 'term one, term two',
-        help: 'Extra words or phrases to flag in your channel. Separate with commas or new lines. Matched even through obfuscation (l33t, look-alike letters).'
-      },
-      {
-        key: 'allow_terms',
-        label: 'Allowed terms',
-        type: 'textarea',
-        placeholder: '',
-        help: 'Words that are fine in your channel: a line containing one is never flagged by the checks above or your blocked terms. Cannot override the safety floor.'
-      }
-    ]
-  },
-  {
-    id: 'shoutout',
-    label: 'Auto Shoutout',
-    tagline: 'Welcome incoming raids with an automatic shoutout.',
-    description:
-      'When another channel raids in, the bot posts a shoutout pointing your chat at the raider. Turn the module on and customize the shoutout line.',
-    icon: 'send',
-    category: 'Community',
-    defaultEnabled: false,
-    replies: [
-      {
-        key: 'shoutout',
-        label: 'Raid shoutout',
-        tagline: 'Automated chat shoutout when raided',
-        event: 'on raid',
-        messageKey: 'message',
-        defaultMessage:
-          'Massive shoutout to {raider} for the raid with {viewers} viewers! Check them out at twitch.tv/{raider.login}',
-        // Tokens the shoutout module resolves (app/sesame/modules/shoutout.go).
-        tokens: ['raider', 'raider.login', 'viewers'],
-        previewSamples: { raider: 'CrustyCrumbs', 'raider.login': 'crustycrumbs', viewers: '42' }
-      }
-    ],
-    settings: [
-      {
-        key: 'native_shoutout',
-        label: 'Also send Twitch shoutout',
-        type: 'toggle',
-        help: "Fires Twitch's own /shoutout on the raider alongside the chat line, which shows their current category and profile card natively. Off by default."
-      }
-    ]
-  },
-  {
-    id: 'alerts',
-    label: 'Chat Alerts',
-    tagline: 'Announce follows, subs, cheers, raids and ad breaks in chat.',
-    description:
-      'The bot posts a chat line when someone follows, subscribes, cheers, or raids, and can announce ad breaks. Turn each alert on or off and customize its message. New alerts default on, except the ad-break alert which stays off until you enable it. A follow alert fires at most once per viewer every three days, so unfollowing and refollowing cannot spam your chat.',
-    icon: 'bell',
-    category: 'Community',
-    defaultEnabled: true,
-    replies: [
-      {
-        key: 'follow',
-        label: 'Follow alert',
-        tagline: 'When someone follows your channel. Fires once per viewer every three days.',
-        event: 'on follow',
-        enableKey: 'followEnabled',
-        messageKey: 'followMessage',
-        defaultMessage: 'Thank you for following the channel, {user}!',
-        // Per-alert tokens/samples mirror the maps in app/sesame/modules/alerts.go.
-        tokens: ['user'],
-        previewSamples: { user: 'sesame_sam' }
-      },
-      {
-        key: 'sub',
-        label: 'Subscribe alert',
-        tagline: 'When someone subscribes or resubscribes.',
-        event: 'on subscribe',
-        enableKey: 'subEnabled',
-        messageKey: 'subMessage',
-        defaultMessage: 'Welcome to the community, {user}! Thank you for subscribing!',
-        tokens: ['user', 'tier'],
-        previewSamples: { user: 'sesame_sam', tier: '1000' }
-      },
-      {
-        key: 'gift',
-        label: 'Gift sub alert',
-        tagline: 'When someone gifts subs. One line per gifter, never per recipient.',
-        event: 'on gift subs',
-        enableKey: 'giftEnabled',
-        messageKey: 'giftMessage',
-        defaultMessage: '{user} just gifted {count} subs to the community! Thank you!',
-        tokens: ['user', 'count', 'tier'],
-        previewSamples: { user: 'GenerousViewer', count: '5', tier: '1000' }
-      },
-      {
-        key: 'cheer',
-        label: 'Cheer alert',
-        tagline: 'When someone cheers bits.',
-        event: 'on cheer',
-        enableKey: 'cheerEnabled',
-        messageKey: 'cheerMessage',
-        defaultMessage: 'Thank you for the {bits} bits, {user}!',
-        tokens: ['user', 'bits'],
-        previewSamples: { user: 'sesame_sam', bits: '500' }
-      },
-      {
-        key: 'raid',
-        label: 'Raid alert',
-        tagline: 'When another channel raids in.',
-        event: 'on raid',
-        enableKey: 'raidEnabled',
-        messageKey: 'raidMessage',
-        defaultMessage: '{user} is raiding the channel with {viewers} viewers! Welcome everyone!',
-        tokens: ['user', 'viewers'],
-        previewSamples: { user: 'CrustyCrumbs', viewers: '42' }
-      },
-      {
-        key: 'ads',
-        label: 'Ad break alert',
-        tagline: 'When an ad break starts. Off by default.',
-        event: 'on ad break',
-        enableKey: 'adsEnabled',
-        messageKey: 'adsMessage',
-        defaultOff: true,
-        defaultMessage: "Ads are rolling for {duration} seconds. Hang tight, we'll be right back!",
-        tokens: ['duration'],
-        previewSamples: { duration: '90' }
-      }
-    ]
-  },
-  // External-stats modules: chat commands answered through the gossip service
-  // (external API proxy + cache). Config keys must match the sesame module
-  // structs (app/sesame/modules/urchin.go, mcsr.go).
-  {
-    id: 'urchin',
-    label: 'Bedwars Stats',
-    tagline: 'Hypixel Bedwars stats, urchin score and blacklist tags in chat.',
-    description:
-      'Viewer commands backed by urchin.gg: daily, weekly and monthly Bedwars sessions, lifetime stats, the Urchin sniper score and active blacklist tags. Commands default to your linked Minecraft account; viewers can also name any player, e.g. "!daily Technoblade".',
-    icon: 'gamepad',
-    category: 'Games',
-    defaultEnabled: false,
-    replies: [
-      {
-        key: 'daily',
-        label: '!daily',
-        tagline: 'Bedwars session since the daily reset.',
-        event: '!daily',
-        command: 'daily',
-        enableKey: 'dailyEnabled',
-        messageKey: 'dailyMessage',
-        defaultMessage: '{player} today: {wins}W {losses}L · {finals} finals · {beds} beds · {fkdr} FKDR',
-        tokens: BW_SESSION_TOKENS,
-        previewSamples: BW_SESSION_SAMPLES
-      },
-      {
-        key: 'weekly',
-        label: '!weekly',
-        tagline: 'Bedwars session since the weekly reset.',
-        event: '!weekly',
-        command: 'weekly',
-        enableKey: 'weeklyEnabled',
-        messageKey: 'weeklyMessage',
-        defaultMessage: '{player} this week: {wins}W {losses}L · {finals} finals · {beds} beds · {fkdr} FKDR',
-        tokens: BW_SESSION_TOKENS,
-        previewSamples: BW_SESSION_SAMPLES
-      },
-      {
-        key: 'monthly',
-        label: '!monthly',
-        tagline: 'Bedwars session since the monthly reset.',
-        event: '!monthly',
-        command: 'monthly',
-        enableKey: 'monthlyEnabled',
-        messageKey: 'monthlyMessage',
-        defaultMessage: '{player} this month: {wins}W {losses}L · {finals} finals · {beds} beds · {fkdr} FKDR',
-        tokens: BW_SESSION_TOKENS,
-        previewSamples: BW_SESSION_SAMPLES
-      },
-      {
-        key: 'stats',
-        label: '!bwstats',
-        tagline: 'Lifetime Bedwars stats.',
-        event: '!bwstats',
-        command: 'bwstats',
-        enableKey: 'statsEnabled',
-        messageKey: 'statsMessage',
-        defaultMessage: '{player}: {stars} stars · {wins} wins · {finals} finals · {fkdr} FKDR · {beds} beds broken',
-        tokens: ['player', 'stars', 'wins', 'losses', 'finals', 'finaldeaths', 'beds', 'fkdr', 'wlr'],
-        previewSamples: {
-          player: 'Technoblade',
-          stars: '402',
-          wins: '1000',
-          losses: '100',
-          finals: '5000',
-          finaldeaths: '500',
-          beds: '2000',
-          fkdr: '10.00',
-          wlr: '10.00'
-        }
-      },
-      {
-        key: 'sniper',
-        label: '!sniper',
-        tagline: 'Urchin (Cubelify overlay) score.',
-        event: '!sniper',
-        command: 'sniper',
-        enableKey: 'sniperEnabled',
-        messageKey: 'sniperMessage',
-        defaultMessage: '{player} urchin score: {score}',
-        tokens: ['player', 'score', 'mode', 'tagcount'],
-        previewSamples: { player: 'Technoblade', score: '7.5', mode: 'warn', tagcount: '1' }
-      },
-      {
-        key: 'tags',
-        label: '!tag',
-        tagline: 'Active Urchin blacklist tags.',
-        event: '!tag',
-        command: 'tag',
-        enableKey: 'tagsEnabled',
-        messageKey: 'tagsMessage',
-        defaultMessage: '{player}: {tags}',
-        tokens: ['player', 'tags', 'tagcount'],
-        previewSamples: {
-          player: 'Technoblade',
-          tags: 'Blatant Cheater (added Jul 3, 2024)',
-          tagcount: '1'
-        }
-      },
-      {
-        key: 'tagdescription',
-        label: '!tagdescription',
-        tagline: 'Blacklist tags with their reasons.',
-        event: '!tagdescription',
-        command: 'tagdescription',
-        enableKey: 'tagDescriptionEnabled',
-        messageKey: 'tagDescriptionMessage',
-        defaultMessage: '{player}: {tags}',
-        tokens: ['player', 'tags', 'tagcount'],
-        previewSamples: {
-          player: 'Technoblade',
-          tags: 'Blatant Cheater (bhop - added Jul 3, 2024)',
-          tagcount: '1'
-        }
-      }
-    ],
-    settings: [
-      {
-        key: 'account',
-        label: 'Linked Minecraft account',
-        type: 'text',
-        placeholder: 'Your Minecraft username',
-        help: 'Default player for every command. Leave blank to use your Twitch username.'
-      }
-    ]
-  },
-  {
-    id: 'mcsr',
-    label: 'MCSR Ranked',
-    tagline: 'Ranked elo and per-stream session stats for MCSR runners.',
-    description:
-      'Viewer commands backed by the MCSR Ranked API: !elo shows the current rating and season record; !session shows elo and wins/losses since the stream started, snapshotting your standing the moment you go live. !elo can name any player (e.g. "!elo Feinberg"); !session always tracks your linked account. !lastmatch shows the most recent match; !record compares two players\' head-to-head record; !lb shows the top 5 of the elo, phase-point or record leaderboard; !race shows the weekly race leader and your own time. !elo, !lastmatch, !record and !lb all accept a trailing "season:11" to look at a past season. !pace, !nethers and !lastfort pull live speedrun splits from PaceMan.gg for the same linked account. !pb shows a personal best: daily/weekly/monthly from PaceMan, or "ranked" for the MCSR Ranked season best; a player name can follow any of those, or stand alone for the all-time PaceMan best.',
-    icon: 'gamepad',
-    category: 'Games',
-    defaultEnabled: false,
-    replies: [
-      {
-        key: 'elo',
-        label: '!elo',
-        tagline: 'Current elo, rank and season record.',
-        event: '!elo',
-        command: 'elo',
-        enableKey: 'eloEnabled',
-        messageKey: 'eloMessage',
-        defaultMessage: '{player}: {elo} elo · rank #{rank} · {wins}W {losses}L this season',
-        tokens: ['player', 'elo', 'rank', 'wins', 'losses', 'matches', 'country'],
-        previewSamples: {
-          player: 'Feinberg',
-          elo: '1650',
-          rank: '12',
-          wins: '40',
-          losses: '20',
-          matches: '61',
-          country: 'us'
-        }
-      },
-      {
-        key: 'session',
-        label: '!session',
-        tagline: 'Elo and record since the stream started.',
-        event: '!session',
-        command: 'session',
-        enableKey: 'sessionEnabled',
-        messageKey: 'sessionMessage',
-        defaultMessage: '{player} this stream: {elochange} elo ({elo} now) · {wins}W {losses}L in {matches} matches',
-        tokens: ['player', 'elo', 'elochange', 'wins', 'losses', 'matches'],
-        previewSamples: {
-          player: 'Feinberg',
-          elo: '1660',
-          elochange: '+24',
-          wins: '3',
-          losses: '1',
-          matches: '4'
-        }
-      },
-      {
-        key: 'pace',
-        label: '!pace',
-        tagline: 'Split averages for the current session, from PaceMan.gg.',
-        event: '!pace',
-        command: 'pace',
-        enableKey: 'paceEnabled',
-        messageKey: 'paceMessage',
-        defaultMessage:
-          '{player} this session: {nethers} nethers (avg {nether}) · bastion {bastion} · fortress {fortress} · fp {firstportal} · {nph} nph',
-        tokens: [
-          'player',
-          'nethers',
-          'nether',
-          'bastion',
-          'fortress',
-          'firststructure',
-          'secondstructure',
-          'firstportal',
-          'stronghold',
-          'end',
-          'finish',
-          'nph'
-        ],
-        previewSamples: {
-          player: 'Feinberg',
-          nethers: '3',
-          nether: '1:42',
-          bastion: '3:55',
-          fortress: '7:12',
-          firststructure: '3:55',
-          secondstructure: '7:12',
-          firstportal: '9:20',
-          stronghold: '12:05',
-          end: '13:50',
-          finish: '0:00',
-          nph: '5.3'
-        }
-      },
-      {
-        key: 'nethers',
-        label: '!nethers',
-        tagline: 'Nether-entrance count and pace for the current session.',
-        event: '!nethers',
-        command: 'nethers',
-        enableKey: 'nethersEnabled',
-        messageKey: 'nethersMessage',
-        defaultMessage: '{player}: {nethers} nethers this session (avg {nether}) · {nph} nph',
-        tokens: ['player', 'nethers', 'nether', 'nph'],
-        previewSamples: { player: 'Feinberg', nethers: '3', nether: '1:42', nph: '5.3' }
-      },
-      {
-        key: 'lastfort',
-        label: '!lastfort',
-        tagline: 'Splits for the most recent run that reached a fortress or bastion.',
-        event: '!lastfort',
-        command: 'lastfort',
-        enableKey: 'lastFortEnabled',
-        messageKey: 'lastFortMessage',
-        defaultMessage:
-          '{player} last fort: nether {nether} · bastion {bastion} · fortress {fortress} · fp {firstportal} · sh {stronghold} · {ago} ago',
-        tokens: [
-          'player',
-          'nether',
-          'bastion',
-          'fortress',
-          'firstportal',
-          'stronghold',
-          'end',
-          'finish',
-          'ago'
-        ],
-        previewSamples: {
-          player: 'Feinberg',
-          nether: '1:42',
-          bastion: '3:55',
-          fortress: '7:12',
-          firstportal: '9:20',
-          stronghold: '—',
-          end: '—',
-          finish: '—',
-          ago: '12m'
-        }
-      },
-      {
-        key: 'lastmatch',
-        label: '!lastmatch',
-        tagline: 'The most recent match: opponent, result, seed and elo change.',
-        event: '!lastmatch',
-        command: 'lastmatch',
-        enableKey: 'lastMatchEnabled',
-        messageKey: 'lastMatchMessage',
-        defaultMessage:
-          '{player} vs {opponent}: {result} · {time} · {seed} {structure} · {elochange} elo · {ago} ago',
-        tokens: ['player', 'opponent', 'result', 'time', 'seed', 'structure', 'elochange', 'ago'],
-        previewSamples: {
-          player: 'Feinberg',
-          opponent: 'lowk3y_',
-          result: 'won',
-          time: '11:03.135',
-          seed: 'Desert Temple',
-          structure: 'Treasure',
-          elochange: '+21',
-          ago: '2m'
-        }
-      },
-      {
-        key: 'record',
-        label: '!record',
-        tagline: 'Head-to-head wins between two players.',
-        event: '!record <player> [player2]',
-        command: 'record',
-        previewArgs: 'Feinberg lowk3y_',
-        enableKey: 'recordEnabled',
-        messageKey: 'recordMessage',
-        defaultMessage: '{playera} {winsa} - {winsb} {playerb} · {played} played',
-        tokens: ['playera', 'playerb', 'winsa', 'winsb', 'played'],
-        previewSamples: { playera: 'Feinberg', playerb: 'lowk3y_', winsa: '20', winsb: '14', played: '34' }
-      },
-      {
-        key: 'lb',
-        label: '!lb',
-        tagline: 'Top 5 of the elo, phase-point or record leaderboard.',
-        event: '!lb [phase [predicted]|record] [country:xx]',
-        command: 'lb',
-        enableKey: 'lbEnabled',
-        messageKey: 'lbMessage',
-        defaultMessage: '{board}: {list}',
-        tokens: ['board', 'list'],
-        previewSamples: { board: 'Elo', list: '#1 Feinberg 2464 · #2 lowk3y_ 2436' }
-      },
-      {
-        key: 'race',
-        label: '!race',
-        tagline: "This week's race leader and the player's own time.",
-        event: '!race',
-        command: 'race',
-        enableKey: 'raceEnabled',
-        messageKey: 'raceMessage',
-        defaultMessage: '#1 {leader} ({leadertime}) · {player}: {time} (#{rank})',
-        tokens: ['leader', 'leadertime', 'player', 'time', 'rank'],
-        previewSamples: { leader: 'gharfyy', leadertime: '2:27.374', player: 'Feinberg', time: '2:40.000', rank: '2' }
-      },
-      {
-        key: 'pb',
-        label: '!pb',
-        tagline: 'Personal best time: PaceMan daily/weekly/monthly/all-time, or the MCSR Ranked season best.',
-        event: '!pb [daily|weekly|monthly|ranked] [player]',
-        command: 'pb',
-        previewArgs: 'daily',
-        enableKey: 'pbEnabled',
-        messageKey: 'pbMessage',
-        defaultMessage: '{player}: {time} ({window} PB)',
-        tokens: ['player', 'time', 'window'],
-        previewSamples: { player: 'Feinberg', time: '6:40.123', window: 'daily' }
-      }
-    ],
-    settings: [
-      {
-        key: 'account',
-        label: 'Linked Minecraft account',
-        type: 'text',
-        placeholder: 'Your Minecraft username',
-        help: 'Default player for every command. Leave blank to use your Twitch username.'
-      }
-    ]
-  },
-  {
-    // !fnstats and !season share one template surface (same tokens, same
-    // sample shape) — FN_STATS_TOKENS/FN_STATS_SAMPLES above are the one
-    // source of truth, mirroring the Bedwars session commands.
-    id: 'fortnite',
-    label: 'Fortnite Stats',
-    tagline: 'Fortnite BR stats and the daily item shop in chat.',
-    description:
-      'One command, four looks: !fn shows a player\'s all-time wins, matches, kills, K/D and win rate with a solo/duo/squad breakdown; !fn season shows the same for the current season (the bot tracks season rollovers automatically); !fn session shows wins, kills and K/D since the stream started, snapshotting your standing the moment you go live; !fn store lists what is in today\'s item shop. The squashed forms !fnstats, !fnseason, !fnsession and !fnstore work too. Link your Epic display name below. Viewers can also name any player, e.g. "!fn Ninja"; !fn session always tracks your linked account. PlayStation and Xbox name lookups are not supported yet.',
-    icon: 'gamepad',
-    category: 'Games',
-    defaultEnabled: false,
-    replies: [
-      {
-        key: 'stats',
-        label: '!fn',
-        tagline: 'All-time Battle Royale stats — !fn, !fn stats or !fnstats.',
-        event: '!fn',
-        command: 'fn',
-        enableKey: 'statsEnabled',
-        messageKey: 'statsMessage',
-        defaultMessage:
-          '{player} all time: {wins} wins in {matches} matches · {winrate}% WR · {kills} kills · {kd} K/D · solo {solowins}W / duo {duowins}W / squad {squadwins}W',
-        tokens: FN_STATS_TOKENS,
-        previewSamples: FN_STATS_SAMPLES
-      },
-      {
-        key: 'season',
-        label: '!fn season',
-        tagline: 'Current-season stats (also !fnseason); season rollovers are tracked automatically.',
-        event: '!fn season',
-        command: 'fn season',
-        enableKey: 'seasonEnabled',
-        messageKey: 'seasonMessage',
-        defaultMessage:
-          '{player} this season: {wins} wins in {matches} matches · {winrate}% WR · {kills} kills · {kd} K/D · solo {solowins}W / duo {duowins}W / squad {squadwins}W',
-        tokens: FN_STATS_TOKENS,
-        previewSamples: {
-          ...FN_STATS_SAMPLES,
-          window: 'season',
-          wins: '10',
-          matches: '21',
-          kills: '163',
-          kd: '14.8',
-          winrate: '47.6',
-          solowins: '4',
-          solomatches: '7',
-          solokd: '12.0',
-          duowins: '3',
-          duomatches: '9',
-          duokd: '9.2',
-          squadwins: '3',
-          squadmatches: '5',
-          squadkd: '21.5'
-        }
-      },
-      {
-        key: 'session',
-        label: '!fn session',
-        tagline: 'Wins, kills and K/D since the stream started (also !fnsession).',
-        event: '!fn session',
-        command: 'fn session',
-        enableKey: 'sessionEnabled',
-        messageKey: 'sessionMessage',
-        defaultMessage:
-          '{player} this stream: {wins} wins in {matches} matches · {winrate}% WR · {kills} kills · {kd} K/D',
-        tokens: FN_SESSION_TOKENS,
-        previewSamples: FN_SESSION_SAMPLES
-      },
-      {
-        key: 'store',
-        label: '!fn store',
-        tagline: "Today's item-shop rotation (also !fnstore).",
-        event: '!fn store',
-        command: 'fn store',
-        enableKey: 'storeEnabled',
-        messageKey: 'storeMessage',
-        defaultMessage: 'Item Shop {date}: {items}',
-        tokens: ['date', 'count', 'items'],
-        previewSamples: {
-          date: '2026-07-09',
-          count: '38',
-          items: 'Peely Bundle (2800), Renegade Raider (1200), Floss (500) +35 more'
-        }
-      }
-    ],
-    settings: [
-      {
-        key: 'account',
-        label: 'Linked account name',
-        type: 'text',
-        placeholder: 'Your Epic display name',
-        help: 'Default player for the stats commands. Leave blank to use your Twitch username.'
-      },
-      {
-        key: 'accountType',
-        label: 'Account platform',
-        type: 'select',
-        placeholder: 'epic',
-        options: [
-          { value: 'epic', label: 'Epic Games' },
-          { value: 'psn', label: 'PlayStation (coming later)' },
-          { value: 'xbl', label: 'Xbox Live (coming later)' }
-        ],
-        help: 'Only Epic display names resolve right now; PlayStation and Xbox lookups come later. Console players: your Epic display name works.'
-      }
-    ]
-  },
-  {
-    // All four !cr views ride gossip's one shared Clash Royale profile cache,
-    // so their token palettes below mirror app/sesame/modules/clashroyale.go.
-    id: 'clashroyale',
-    label: 'Clash Royale Stats',
-    tagline: 'Clash Royale profiles, decks and Path of Legends standing in chat.',
-    description:
-      'One command, four looks: !cr shows a player\'s lifetime profile (level, win/loss record, win rate, three-crown wins, clan); !cr decks lists their current battle deck with the average elixir cost; !cr ranked shows their Path of Legends standing (falling back to legacy league seasons for older accounts); !cr road shows their trophy-road record and arena. The squashed forms !crstats, !crdecks, !crranked and !crroad work too. Link your player tag below — Clash Royale has no name lookup, so a tag like #P2LQ0GR is required. Viewers can also name any tag, e.g. "!cr #P2LQ0GR".',
-    icon: 'gamepad',
-    category: 'Games',
-    defaultEnabled: false,
-    replies: [
-      {
-        key: 'stats',
-        label: '!cr',
-        tagline: 'Lifetime profile — !cr, !cr stats or !crstats.',
-        event: '!cr',
-        command: 'cr',
-        enableKey: 'statsEnabled',
-        messageKey: 'statsMessage',
-        defaultMessage: '{player} · level {level} · {wins}W/{losses}L · {winrate}% WR · {crowns} three-crowns · {clan}',
-        tokens: ['player', 'tag', 'level', 'wins', 'losses', 'draws', 'battles', 'winrate', 'crowns', 'challengemax', 'donations', 'totaldonations', 'clan', 'favcard'],
-        previewSamples: {
-          player: 'Bagel', tag: '#P2LQ0GR', level: '62', wins: '600', losses: '300', draws: '100',
-          battles: '1000', winrate: '60', crowns: '120', challengemax: '12', donations: '50',
-          totaldonations: '10000', clan: 'Bakery', favcard: 'Knight'
-        }
-      },
-      {
-        key: 'decks',
-        label: '!cr decks',
-        tagline: "The current battle deck with its average elixir (also !crdecks).",
-        event: '!cr decks',
-        command: 'cr decks',
-        enableKey: 'decksEnabled',
-        messageKey: 'decksMessage',
-        defaultMessage: "{player}'s deck ({count}/8): {cards} · avg elixir {elixir}",
-        tokens: ['player', 'tag', 'cards', 'support', 'elixir', 'count'],
-        previewSamples: {
-          player: 'Bagel', tag: '#P2LQ0GR',
-          cards: 'Knight, Archers, Goblins, Giant, P.E.K.K.A, Minions, Fireball, Cannon',
-          support: 'Tower Troop', elixir: '3.75', count: '8'
-        }
-      },
-      {
-        key: 'ranked',
-        label: '!cr ranked',
-        tagline: 'Path of Legends standing (also !crranked); legacy league seasons fill in for older records.',
-        event: '!cr ranked',
-        command: 'cr ranked',
-        enableKey: 'rankedEnabled',
-        messageKey: 'rankedMessage',
-        defaultMessage: '{player} Path of Legends: league {league} · {trophies} trophies · rank #{rank} · best {besttrophies}',
-        tokens: ['player', 'tag', 'league', 'trophies', 'rank', 'prevleague', 'prevtrophies', 'bestleague', 'besttrophies', 'bestrank'],
-        previewSamples: {
-          player: 'Bagel', tag: '#P2LQ0GR', league: '10', trophies: '2100', rank: '321',
-          prevleague: '10', prevtrophies: '2050', bestleague: '10', besttrophies: '2400', bestrank: '42'
-        }
-      },
-      {
-        key: 'road',
-        label: '!cr road',
-        tagline: 'Trophy-road record and arena (also !crroad).',
-        event: '!cr road',
-        command: 'cr road',
-        enableKey: 'roadEnabled',
-        messageKey: 'roadMessage',
-        defaultMessage: '{player}: {trophies} trophies · best {besttrophies} · {arena}',
-        tokens: ['player', 'tag', 'trophies', 'besttrophies', 'arena'],
-        previewSamples: { player: 'Bagel', tag: '#P2LQ0GR', trophies: '9123', besttrophies: '9345', arena: 'Legendary Arena' }
-      }
-    ],
-    settings: [
-      {
-        key: 'account',
-        label: 'Linked player tag',
-        type: 'text',
-        placeholder: '#P2LQ0GR',
-        help: 'Default player for every command. Clash Royale has no name lookup, so this must be a player tag; leave blank only if you always type one.'
-      }
-    ]
-  },
-  {
-    // All five !val views are gossip lookups, so their token palettes below
-    // mirror app/sesame/modules/valorant.go.
-    id: 'valorant',
-    label: 'Valorant Stats',
-    tagline: 'Valorant ranks, match history, leaderboards and the daily shop rotation in chat.',
-    description:
-      'One command, five looks: !val shows a player\'s competitive standing (tier, RR, last game\'s RR change, peak tier); !val matches lists their recent ranked games as agent K/D/A results; !val account shows who a Riot ID resolves to, with their account level; !val lb prints the top 10 of any regional leaderboard, PC or console; !val shop shows today\'s global skin rotation with VP prices and the reset countdown. The squashed forms !valrank, !valmatches, !valaccount, !vallb and !valshop work too. Link your Riot ID below ("Name#Tag") — viewers can also name any player, e.g. "!val Frosty#EUW1", and add a shard or ladder anywhere in the line to scope that one lookup ("!val eu Frosty#EUW1", "!val lb console ap"). Leave the region blank and the bot detects the shard from the account itself.',
-    icon: 'gamepad',
-    category: 'Games',
-    defaultEnabled: false,
-    replies: [
-      {
-        key: 'rank',
-        label: '!val',
-        tagline: 'Competitive standing — !val, !val rank or !valrank.',
-        event: '!val',
-        command: 'val',
-        enableKey: 'rankEnabled',
-        messageKey: 'rankMessage',
-        defaultMessage: '{player} · {tier} · {rr} RR ({lastchange}) · peak {peaktier}',
-        tokens: ['player', 'region', 'tier', 'elo', 'rr', 'lastchange', 'peaktier', 'placement'],
-        previewSamples: {
-          player: 'Frosty#EUW1', region: 'eu', tier: 'Immortal 2', elo: '1832', rr: '67',
-          lastchange: '-12', peaktier: 'Immortal 1', placement: '513'
-        }
-      },
-      {
-        key: 'matches',
-        label: '!val matches',
-        tagline: 'Recent ranked games (also !valmatches).',
-        event: '!val matches',
-        command: 'val matches',
-        enableKey: 'matchesEnabled',
-        messageKey: 'matchesMessage',
-        defaultMessage: "{player}'s last {count}: {matches}",
-        tokens: ['player', 'region', 'count', 'matches', 'lastago'],
-        previewSamples: {
-          player: 'Frosty#EUW1', region: 'eu', count: '2',
-          matches: 'Jett 20/14/7 win on Haven, Omen 9/17/3 loss on Ascent', lastago: '2h ago'
-        }
-      },
-      {
-        key: 'account',
-        label: '!val account',
-        tagline: 'Who an ID resolves to, with account level (also !valaccount).',
-        event: '!val account',
-        command: 'val account',
-        enableKey: 'accountEnabled',
-        messageKey: 'accountMessage',
-        defaultMessage: '{player} · account level {level}',
-        tokens: ['player', 'puuid', 'region', 'level', 'card', 'title'],
-        previewSamples: {
-          player: 'Frosty#EUW1', puuid: 'a1b2c3d4-05e6-47f8-89a0-b1c2d3e4f5a6',
-          region: 'eu', level: '142', card: 'Silver Card', title: 'Radiant'
-        }
-      },
-      {
-        key: 'board',
-        label: '!val lb',
-        tagline: 'Regional top 10, PC or console (also !vallb).',
-        event: '!val lb',
-        command: 'val lb',
-        enableKey: 'boardEnabled',
-        messageKey: 'boardMessage',
-        defaultMessage: '{board}: {entries}',
-        tokens: ['player', 'board', 'count', 'entries'],
-        previewSamples: {
-          player: 'Frosty#EUW1', board: 'ap/console', count: '2',
-          entries: '#4 Zekken#5221 (431 RR), #5 Frosty#EUW1 (402 RR)'
-        }
-      },
-      {
-        key: 'shop',
-        label: '!val shop',
-        tagline: "Today's global skin rotation with reset countdown (also !valshop).",
-        event: '!val shop',
-        command: 'val shop',
-        enableKey: 'shopEnabled',
-        messageKey: 'shopMessage',
-        defaultMessage: 'Daily rotation ({count}): {items} · resets in {reset}',
-        tokens: ['count', 'items', 'reset'],
-        previewSamples: {
-          count: '2', items: 'Reaver Vandal (1775 VP), Ion Frenzy (875 VP)', reset: '2h 30m'
-        }
-      }
-    ],
-    settings: [
-      {
-        key: 'account',
-        label: 'Linked Riot ID',
-        type: 'text',
-        placeholder: 'Frosty#EUW1',
-        help: 'Default player for every command, as "Name#Tag" — Valorant has no name-only lookup, so the #tag is required.'
-      },
-      {
-        key: 'region',
-        label: 'Region shard',
-        type: 'text',
-        placeholder: 'eu',
-        help: 'na, eu, ap, kr, br or latam. Leave blank and the bot detects the shard from the account itself.'
-      },
-      {
-        key: 'platform',
-        label: 'Ladder',
-        type: 'select',
-        placeholder: 'pc',
-        options: [
-          { value: 'pc', label: 'PC' },
-          { value: 'console', label: 'Console' }
-        ],
-        help: 'Ranks and leaderboards are tracked as separate ladders per platform; blank means PC.'
-      }
-    ]
-  },
-  {
-    id: 'queue',
-    label: 'Play Queue',
-    tagline: 'Let viewers line up to play with you, first come first served.',
-    description:
-      'Viewers type !join to get in line and !list to see who is next (the first 10). You (and your mods) run the line from chat: !queue open and !queue close accept or stop new joins, !queue next pulls up the next player, !queue remove <user> takes someone out, and !queue clear empties it. Viewers can step out any time with !leave. Turn the module on to enable the commands; the line survives closing so you can play through everyone already waiting.',
-    icon: 'list',
-    category: 'Community',
-    defaultEnabled: false,
-    // The conversational replies are customizable per broadcaster; the roster
-    // (!list), the status readout and the system/error lines stay fixed (see
-    // app/sesame/modules/queue.go). The command list below is read-only. Each
-    // reply rehearses as its command (a viewer types the trigger, the bot
-    // answers) with this reply's own sample values.
-    replies: [
-      {
-        key: 'join',
-        label: 'Join confirmation',
-        tagline: 'When a viewer joins the line.',
-        event: '!join',
-        command: 'join',
-        messageKey: 'joinMessage',
-        defaultMessage: '@{user} you joined the queue at position #{pos}.',
-        tokens: ['user', 'pos'],
-        previewSamples: { user: 'sesame_sam', pos: '3' }
-      },
-      {
-        key: 'already',
-        label: 'Already in queue',
-        tagline: 'When a viewer who is already in line types !join again.',
-        event: '!join',
-        command: 'join',
-        messageKey: 'alreadyMessage',
-        defaultMessage: '@{user} you are already in the queue at position #{pos}.',
-        tokens: ['user', 'pos'],
-        previewSamples: { user: 'sesame_sam', pos: '2' }
-      },
-      {
-        key: 'leave',
-        label: 'Leave confirmation',
-        tagline: 'When a viewer steps out of the line.',
-        event: '!leave',
-        command: 'leave',
-        messageKey: 'leaveMessage',
-        defaultMessage: '@{user} you left the queue.',
-        tokens: ['user'],
-        previewSamples: { user: 'sesame_sam' }
-      },
-      {
-        key: 'next',
-        label: 'Next player up',
-        tagline: 'When you pull up the next player.',
-        event: '!queue next',
-        command: 'queue next',
-        messageKey: 'nextMessage',
-        defaultMessage: '@{target} you are up next! ({count} still waiting)',
-        tokens: ['target', 'count'],
-        previewSamples: { target: 'ferret_king', count: '2' }
-      },
-      {
-        key: 'opened',
-        label: 'Queue opened',
-        tagline: 'Announced when you open the queue.',
-        event: '!queue open',
-        command: 'queue open',
-        messageKey: 'openedMessage',
-        defaultMessage: 'The queue is now open! Type !join to get in line.',
-        previewSamples: {}
-      },
-      {
-        key: 'closed',
-        label: 'Queue closed',
-        tagline: 'Announced when you close the queue.',
-        event: '!queue close',
-        command: 'queue close',
-        messageKey: 'closedMessage',
-        defaultMessage: 'The queue is now closed to new joins.',
-        previewSamples: {}
-      }
-    ],
-    commands: [
-      { trigger: '!join', summary: 'Get in line to play (also !queue join).' },
-      { trigger: '!leave', summary: 'Step out of the line (also !queue leave).' },
-      { trigger: '!list', summary: 'Show the next 10 players waiting (also !queue list).' },
-      { trigger: '!queue', summary: 'Show whether the queue is open and how many are waiting.' },
-      { trigger: '!queue open', summary: 'Start accepting joins.', perm: 'mod' },
-      { trigger: '!queue close', summary: 'Stop accepting joins; the line is kept.', perm: 'mod' },
-      { trigger: '!queue next', summary: 'Pull up the next player and announce them.', perm: 'mod' },
-      { trigger: '!queue remove <user>', summary: 'Take someone out of the line.', perm: 'mod' },
-      { trigger: '!queue clear', summary: 'Empty the line.', perm: 'mod' }
-    ]
-  },
-  {
-    id: 'raffle',
-    label: 'Raffle',
-    tagline: 'Timed random draws your chat enters with !join.',
-    description:
-      'Open a raffle and viewers type !join to enter. While it runs the bot posts a time-left reminder every few minutes (the cadence is yours to set), and when time runs out it draws automatically — winners are picked uniformly at random from everyone who entered, every entry counts once, and the draw leaves a verifiable receipt behind. Winners confirm with !claim inside a 15-minute window. You (and your mods) also control everything from chat: !raffle open starts one, !raffle draw closes early and announces, !raffle cancel tears it down without drawing. When both this and the Play Queue are on, !join belongs to the raffle and the queue is reachable through !queue join.',
-    icon: 'gem',
-    category: 'Community',
-    defaultEnabled: false,
-    // The viewer-facing conversational replies are customizable per broadcaster
-    // and each rehearses as its command (a viewer types the trigger, the bot
-    // answers) with this reply's own sample values. The status readout, mod
-    // confirmations and claim outcomes beyond the first stay fixed system text
-    // (see app/sesame/modules/raffle.go); so do the engine-posted auto-close
-    // and reminder announcements.
-    replies: [
-      {
-        key: 'joined',
-        label: 'Entry confirmation',
-        tagline: 'When a viewer joins the raffle.',
-        event: '!join',
-        command: 'join',
-        messageKey: 'joinMessage',
-        defaultMessage: "@{user} you're in! {count} entered so far. Good luck!",
-        tokens: ['user', 'count'],
-        previewSamples: { user: 'sesame_sam', count: '12' }
-      },
-      {
-        key: 'already',
-        label: 'Already entered',
-        tagline: 'When a viewer who is already in types !join again.',
-        event: '!join',
-        command: 'join',
-        messageKey: 'alreadyMessage',
-        defaultMessage: '@{user} you are already in this raffle ({count} entered).',
-        tokens: ['user', 'count'],
-        previewSamples: { user: 'sesame_sam', count: '13' }
-      },
-      {
-        key: 'noRaffle',
-        label: 'No raffle running',
-        tagline: "When someone joins while no raffle is open.",
-        event: '!join',
-        command: 'join',
-        messageKey: 'noRaffleMessage',
-        defaultMessage: '@{user} no raffle is running right now.',
-        tokens: ['user'],
-        previewSamples: { user: 'sesame_sam' }
-      },
-      {
-        key: 'opened',
-        label: 'Raffle opened',
-        tagline: 'When a raffle starts and entries open.',
-        event: '!raffle open',
-        command: 'raffle open',
-        messageKey: 'openedMessage',
-        defaultMessage: 'Raffle is LIVE! Type !join to enter — drawing in {mins} min!',
-        tokens: ['mins'],
-        previewSamples: { mins: '10' }
-      },
-      {
-        key: 'won',
-        label: 'Winners announced',
-        tagline: 'When a draw names its winners (manual or timed).',
-        event: '!raffle draw',
-        command: 'raffle draw',
-        messageKey: 'wonMessage',
-        defaultMessage:
-          '{targets} — congratulations! You won the raffle ({count} winner(s) from {entrants})! Type !claim within {claim} min to confirm your prize!',
-        tokens: ['targets', 'count', 'entrants', 'claim'],
-        previewSamples: { targets: '@maya_live, @crustycrumbs', count: '2', entrants: '18', claim: '15' }
-      },
-      {
-        key: 'claimOk',
-        label: 'Prize confirmed',
-        tagline: 'When a winner confirms with !claim.',
-        event: '!claim',
-        command: 'claim',
-        messageKey: 'claimOkMessage',
-        defaultMessage: '@{user} your prize is confirmed — enjoy!',
-        tokens: ['user'],
-        previewSamples: { user: 'maya_live' }
-      }
-    ],
-    commands: [
-      { trigger: '!winner', summary: "Recall the last draw's winners and how many confirmed." },
-      { trigger: '!raffle', summary: 'Show whether a raffle is running, entries and time left.' },
-      { trigger: '!raffle open [minutes] [winners] [remind]', summary: 'Start a raffle (defaults 10 min, 1 winner, reminders every 5 min; remind 0 turns reminders off).', perm: 'mod' },
-      { trigger: '!raffle draw [winners]', summary: 'Close now and announce the winners.', perm: 'mod' },
-      { trigger: '!raffle close', summary: 'Same as draw with the configured winner count.', perm: 'mod' },
-      { trigger: '!raffle cancel', summary: 'Tear down without drawing.', perm: 'mod' }
-    ]
-  },
-  {
-    id: 'quotes',
-    label: 'Quotes',
-    tagline: 'Save the best things said on stream and replay them in chat.',
-    description:
-      'Keep a channel quote book. Someone saves a line with !addquote the text (or !quote "the text") and the bot numbers it and stamps the save date. Anyone replays one with !quote for a random pick, !quote 12 for a specific number, or !quote ferret for a random quote containing that word — the bot answers "Quote #12: the text (2026-07-11)". Choose who is allowed to save and who may rewrite one (!quote edit 12 the fixed text) below; both default to moderators. Mods remove a mistake with !quote remove 12; numbers are never reused, so an old number keeps pointing at the same quote. The replies are fixed system text (localized to your console language).',
-    icon: 'quote',
-    category: 'Community',
-    defaultEnabled: false,
-    // Bespoke page: the quote book (list + add/edit/remove) plus the enable
-    // and save/edit permission settings live on /quotes, not the generic
-    // reply page.
-    href: '/quotes',
-    // Like timers, the quote book also rides the 'commands' grant.
-    delegateSections: ['modules', 'commands'],
-    replies: [],
-    commands: [
-      { trigger: '!quote', summary: 'Post a random saved quote (also !quotes).' },
-      { trigger: '!quote <number>', summary: 'Post that exact quote.' },
-      { trigger: '!quote <word>', summary: 'Post a random quote containing that word.' },
-      { trigger: '!addquote <text>', summary: 'Save a new quote (also !quoteadd or !quote "text"; who can save is set above); the save date is kept with it.' },
-      { trigger: '!quote edit <number> <text>', summary: 'Rewrite a saved quote in place (who can edit is set above); its number and date stay.' },
-      { trigger: '!quote remove <number>', summary: 'Delete a quote; its number is retired.', perm: 'mod' }
-    ]
-  },
-  {
-    id: 'govee',
-    label: 'Govee Lights',
-    tagline: 'Let viewers recolour your Govee lights with channel points.',
-    description:
-      'Viewers redeem a channel-points reward, type a colour (a name like "blue" or a hex like #00ccff), and the bot turns your Govee light on and sets it. Live only: redemptions off-stream are refunded automatically. To get your Govee API key: open the Govee Home app, tap Profile (bottom right), tap the settings gear (top right), tap "Apply for API Key", fill in the short form, and Govee emails you a key within a few minutes. Then on this page: paste the key (we store it encrypted and never show it back), pick the light to control, and create the reward.',
-    icon: 'power',
-    category: 'Community',
-    defaultEnabled: false,
-    // The generic reply page cannot express key custody + a device picker, so
-    // the tile opens a bespoke inspector instead.
-    href: '/govee',
-    replies: []
-  }
-];
 
 // GOVEE_COLOR_NAMES are the colour words a viewer may type in the Govee reward
 // input. It mirrors the sesame colour parser's named palette
